Makes consents a nested resource
